Implement the clean-cache cronjob
The clean-cache-ipi.sh script will run to evict the nodes from having old and non-used boot images in the cache. We set the cron to run daily.
The script, instead, checks the last access time and compares it with the current date, deleting files that are not accessed for more than 15 days.