Implement the reserve-hosts.sh
The reserve-hosts.sh is a concurrency-safe script allowing a user to request a given number of master and worker nodes from the CI pool defined in the RESERVED_FILE as a csv.
Based on the kind of installation that the reservation is meant for, a user may request vIPs to use for the ingress and the APIs. Those IPs are chosen from two different subnets than the ones that are meant for the hosts IP Address Management.
The output of the script, reporting the hosts being reserved with their hardware and job-specific information, is stored in csv and yaml files, to be consumed by the CI jobs.