Add pdu reboot script in bastion (#94)
pdu reboot has been failing for a long time as the PDU ssh server does
not support latest algorithims. It works from bastion with the
`-o 'PubkeyAcceptedKeyTypes=+ssh-rsa' -o 'HostKeyAlgorithms=+ssh-rsa'`
params set on it. This is the best workaround to make sure we have the
pdu reboot running successfully while we look into options to update the
pdu servers.